+++ flows/landing_zone_delete.py
@@ -1,7 +1,7 @@
 from config import settings
 
 from .base_flow import BaseLinearFlow
-from apis.irods_utils import get_project_path
+from apis.irods_utils import get_landing_zone_path
 from tasks import omics_tasks, irods_tasks
 
 
@@ -13,10 +13,11 @@ class Flow(BaseLinearFlow):
 
     def validate(self):
         self.required_fields = [
-            'zone_uuid',
             'zone_title',
-            'user_name',
-            'user_uuid']
+            'zone_uuid',
+            'study_uuid',
+            'assay_uuid',
+            'user_name']
         return super(Flow, self).validate()
 
     def build(self, force_fail=False):
@@ -25,10 +26,12 @@ class Flow(BaseLinearFlow):
         # Setup
         ########
 
-        project_path = get_project_path(self.project_uuid)
-        zone_path = project_path + '/landing_zones/' + \
-            self.flow_data['user_name'] + '/' + \
-            self.flow_data['zone_title']
+        zone_path = get_landing_zone_path(
+            project_uuid=self.project_uuid,
+            user_name=self.flow_data['user_name'],
+            study_uuid=self.flow_data['study_uuid'],
+            assay_uuid=self.flow_data['assay_uuid'],
+            zone_title=self.flow_data['zone_title'])
 
         ##############
         # iRODS Tasks